+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # v6.6.0
 ## New
 - Updated starter media catalog
+- Added more public cocktail filters
 
 ## Fixes
 - Fixed author and year not being imported via datapack